feat: Add ElectraRemote class for controlling Electra AC
This commit adds a new header file `electraRemote.h` and its corresponding implementation file. The `ElectraRemote` class provides methods for setting the fan speed, temperature, power state, and temperature mode of an Electra AC unit. It also includes private helper methods for transmitting the current state and IR commands.

The purpose of this change is to add support for controlling Electra AC units using IR signals. This functionality will be used in the final project.